refactor: Move most interfaces out of the JSX namespace
